correct smells
