+++ projeto-intermedio/aeolus/src/main/java/lei/tqs/aeolus/external_api/open_weather_utils/OpenWeatherData.java
@@ -1,4 +1,4 @@
-package lei.tqs.aeolus.externalAPI.OpenWeatherUtils;
+package lei.tqs.aeolus.external_api.open_weather_utils;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.*;
@@ -9,8 +9,7 @@ import lombok.*;
 @AllArgsConstructor
 @ToString
 @JsonIgnoreProperties(ignoreUnknown = true)
-
-public class Coordinate {
-    private String lon;
-    private String lat;
+public class OpenWeatherData {
+    private OpenWeatherMeasure components;
+    private long dt;
 }
