--- projeto-intermedio/aeolus/src/test/java/lei/tqs/aeolus/AeolusApplicationTests.java
@@ -1,13 +1,0 @@
-package lei.tqs.aeolus;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AeolusApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
