+++ projeto-intermedio/aeolus/src/test/java/lei/tqs/aeolus/cache/SystemClock.java
@@ -1,9 +1,6 @@
 package lei.tqs.aeolus.cache;
 
 public class SystemClock {
-    public void sleep(long sec) throws InterruptedException {
-        Thread.sleep(sec * 1000);
-    }
 
     public static void whileSleep(long sec) {
         long current = System.currentTimeMillis();
